+++ packages/cubejs-schema-compiler/test/integration/postgres/sql-generation.test.ts
@@ -277,7 +277,7 @@ describe('SQL Generation', () => {
           longitude: { sql: \`longitude\` }
         },
         questionMark: {
-          sql: \`replace('some string question string???', 'string', 'with some ???')\`,
+          sql: \`replace('some string question string ? ?? ???', 'string', 'with some ? ?? ???')\`,
           type: \`string\`
         }
       }