+++ rust/cubesqlplanner/cubesqlplanner/src/plan/builder.rs
@@ -1,3 +1,0 @@
-pub fn build_sql_query_and_params() -> String {
-    "SELECT".to_string()
-}