+++ rust/cubesqlplanner/cubesqlplanner/src/planner/sql_evaluator/symbols/mod.rs
@@ -1,17 +1,17 @@
 mod cube_symbol;
 mod dimension_symbol;
-mod join_symbol;
 mod measure_symbol;
 mod member_symbol;
 mod member_symbol_type;
 mod simple_sql;
+mod symbol_factory;
 
 pub use cube_symbol::{
     CubeNameSymbol, CubeNameSymbolFactory, CubeTableSymbol, CubeTableSymbolFactory,
 };
 pub use dimension_symbol::{DimensionSymbol, DimensionSymbolFactory};
-pub use join_symbol::{JoinConditionSymbol, JoinConditionSymbolFactory};
 pub use measure_symbol::{MeasureSymbol, MeasureSymbolFactory};
-pub use member_symbol::{MemberSymbol, MemberSymbolFactory};
+pub use member_symbol::MemberSymbol;
 pub use member_symbol_type::MemberSymbolType;
 pub use simple_sql::{SimpleSqlSymbol, SimpleSqlSymbolFactory};
+pub use symbol_factory::SymbolFactory;